+++ CHANGES.md
@@ -3,5 +3,11 @@ Changelog
 
 For more information, check the README.md.
 
+0.1.0, 2016/06/13 -- Inclusion of two sample hidden services: hello-onion and headers-onion.
+- Inclusion of two sample hidden services: hello-onion and headers-onion.
+- Abstraction of the process of creating new services and making them available through Tor.
+- Added the possibility of configuring where we want the hidden services to be run.
+- Abstraction of the parameter configuration.
+
 0.0.1, 2016/06/13 -- Initial commit.
 - Initial commit.